Updating the platform
