+++ shared-ui/src/jsMain/kotlin/co.touchlab.droidcon/ui/util/NavigationBackPressWrapper.js.kt
@@ -4,6 +4,6 @@ import androidx.compose.runtime.Composable
 
 @Composable
 internal actual fun NavigationBackPressWrapper(content: @Composable (() -> Unit)) {
-    // For now no back press wrapping is needed on Android.
+    // No back-press handling needed on web.
     content()
 }
