+++ web/src/jsMain/composeResources/drawable/compose-multiplatform.xml
