+++ web/src/jsMain/resources/favicon.png
