+++ web/src/jsMain/resources/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>KotlinProject</title>
+    <title>Droidcon</title>
     <link type="text/css" rel="stylesheet" href="styles.css">
     <link rel="icon" href="favicon.png" type="image/x-icon">
 </head>
