--- web/src/wasmJsMain/kotlin/org/example/project/Platform.wasmJs.kt
@@ -1,7 +1,0 @@
-package org.example.project
-
-class WasmPlatform : Platform {
-    override val name: String = "Web with Kotlin/Wasm"
-}
-
-actual fun getPlatform(): Platform = WasmPlatform()
